backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
python-decouple>=3.8
cloudinary>=1.36
django-cloudinary-storage>=0.3.0
psycopg2-binary>=2.9
Pillow>=10.0
django-filter>=23.0
drf-spectacular>=0.27
gunicorn>=21.0
